Add loading feedback on MatchDetail page
A loading feedback has been added while fetching, editing, and deleting a match on the MatchDetail page.
By introducing a new boolean _loading, spinner will be shown or hidden based on the state of _loading. This change is necessary to provide a better user experience and inform users that an action is being processed.