Add functionality to edit match results directly from result display
This commit introduces the ability for match creators to edit match results directly from the result display on the match detail page. Previously, match results could only be modified from the list of player details, which was a bit unintuitive and buried this functionality. With this update, the match results progressively reveal the option to edit if you are the creator of the match. This leads to a better user experience by making this functionality more discoverable.