"Add match result confirmation buttons for both teams"
This commit makes visual changes to the MatchDetail, ShowResults, and AddResults razor component files. The key change is that buttons have been added to allow both teams to independently confirm the result of a match, thus ensuring fair play and double checking. This was done inside of `ShowResults.razor` conditional rendering. Furthermore, a code refactor for the `AddResults.razor` was performed to simplify the procedure of confirming the results, improving readability and maintainability.