Add match result confirmation functionality
This commit introduces new functionality to the MatchController that allows users to confirm the result of their match. An HTTP endpoint, [HttpPost("{matchId:guid}/confirm-result/{confirmation:bool}")], has been added that saves the confirmation of a team's result to the Match instance in the database. The function takes the match id and confirmation as parameters from the route and the user id from the User token. In the case where the user confirms the result, the match's corresponding score confirmation for the user's team is set to the confirmation parameter. These changes were introduced to give users the ability to verify the outcome of their matches and preserve the integrity of the results.

The front-end ShowResult component has also been modified. Two new modals were added for the user to either confirm or delete their confirmation of the result. The div containing the confirmation buttons was also updated. Buttons to confirm or delete the result confirmation are now conditionally displayed based on whether or not the user's team has already confirmed the result. These front-end updates ensure users can interact with the new server-side match result confirmation functionality conveniently.