Add score confirmation check to match controller and components
A check for whether the scores of both teams have been confirmed has been added. In the MatchController, a conflict is returned when trying to fetch a game that has already been confirmed by both teams. In the ShowResults and MatchDetailPlayers components, buttons for confirming and deleting confirmation of results are now only displayed if the results have not yet been confirmed by both teams. Additionally, teams with confirmed scores are displayed in green to indicate their confirmed status. This change is made to prevent potential data inconsistencies and further unnecessary operations once the match scores have been confirmed.