Add 'Finished' indicator to matches
A 'Finished' field has been added to the MatchController, AllMatches.razor view and MatchResponse to give an immediate visual indication of a match's completion status. If both teams have confirmed their scores, the match is considered 'Finished'. A green 'FINISHED' stamp will be displayed on the match card in the 'AllMatches' view.