Add notification for open slots in matches
The commit introduces a "Open slots" stamp on the match card interface for games where the players count is less than 4. A new CSS class for this stamp has also been added for styling purposes, functioning similarly to the existing "finished" stamp. This provides a visual idea for users about the matches still requiring players.