Update scoring UI layout in AddResults component
The scoring input layout in the AddResults razor component was updated to improve UX. The foreach loop was replaced with a for loop to access and display scores. Also, form control classes were appended with 'set-score' to provide targeted styling. Justification, position and margin styles were adjusted to centralize elements and provide spacing for clarity. Input width and margin were additionally set for '.set-score' class.