Update UI elements in MatchDetail page
This commit replaces some HTML div elements with Razor components in the MatchDetail page - for better reusability and cleaner code. The separation div is replaced with a Divider component, making the code more semantic. Also, the display of match location is commented out for now, as it might not be necessary and can be confusing.