Refactored confirm result buttons and modals in ShowResults
Adjusted the UI and logical flow of the confirm/deny result functionality in the ShowResults component. The main changes here are to the buttons for confirming results and deleting confirmations. The standard HTML buttons have been replaced with Blazorise buttons in the size "Small" and color "Success" (for confirming result) and "Danger" (for deleting confirmation), providing a clearer visual indication of each button's purpose.

Also, the prompts to confirm these actions are now handled directly in the buttons' click functions with the Confirm method from the MessageService, which was added to the component's dependency injections, rather than through separate modals invoked elsewhere in the code. This simplifies the code structure and the page layout, as it no longer needs separate modals for each type of confirmation.

Lastly, the modal to confirm deletion of a previously confirmed result has been removed altogether because its functionality is now covered by the Confirm method in the DeleteConfirmResult function.