Add dividers, refactor buttons, remove modal in ShowResults
A divider has been added above the buttons in ShowResults.razor for better separation of sections. The design of the buttons has been refactored for uniformity in the application: size and colors are now defined using predefined categories, and outlining is activated. "Delete score" button now uses a confirmation message directly, eliminating the need for a separate confirmation modal, thus simplifying the component's code. This makes the code cleaner and improves the overall user experience.