Add Alpha version announcement for users
Added a new feature to address users about the Alpha version of the application in a more personalized way. Two new public methods have been introduced to PlayerController.cs to obtain and modify the status of the development announcement. They work alongside the new property DevelopmentAnnouncementReadIt, ensuring that users won't repeatedly see the announcement. The relevant UI components were also added in Index.razor and DevelopmentVersionAnnouncement.razor to display this information to the users.