Add check for removing players from scored match
This commit introduces an additional validation for removing players from a match. Now, if a match score has been confirmed either for TeamOne or TeamTwo, users are prevented from removing any players. This is to ensure the consistency and accuracy of a match's final score, as it could be disrupted if players were allowed to be removed.