Reduce username length and hide email field
Updated the maximum length for the username field in the ExternalLogin model from 14 to 10 characters. This is to ensure username consistency across all platforms. Also, the email input field in the ExternalLogin view is now hidden from the user. This is in response to recent changes, as email is no longer required for external login.