Update external login user prompt text
The user prompt text under external login section has been modified. The previous text asked users to enter their username and confirm the email account. The new text removes the aspect of email confirmation and only asks the user to enter their username. This is to streamline the user login flow and reduce redundancy as the email confirmation is already handled in another process step.