Update User Identification & Initial ELO score Method; Disable Certain Account Settings
User identification in "MatchController" was updated to be specifically sorted by a unique User ID, not by an unclear "UserId". This guarantees that there is no conflict or uncertainty when assigning matches to users.

In the "ConfirmEmail" confirmation process, the assignment of initial ELO score for new users was updated. Instead of assigning a static value of 0, a new EloHistory object is created with an initial score of 1500. This offers a more realistic starting point for new players.

In the "_ManageNav" section, links to "Two-factor authentication" and "Personal data" were temporarily disabled for review and debugging. This is a temporary change and these features will be re-enabled after testing.