Add match winner calculation and scoring visualization
Implemented a feature that calculates the match winner based on the set wins in the 'MatchController.cs'. If there is no winner, a null value will be returned. Additionally, the Player query in the 'MatchController.cs' is adjusted to filter by the current user's id instead of a static id. In the frontend, changes were made in 'ShowResults.razor' to visually distinguish between team scores by assigning green color to the winning score and red to the losing score. To store the match winner data, a 'TeamWinner' field has been added in 'MatchResponse.cs'. Lastly, corrected format error in 'Startup.cs' by removing redundant spaces after blobStorageSection method in ConfigureServices.