Add initial PreviousElo to EloHistory for new players
Previously, a new user was only assigned a CurrentElo upon creation. This patch includes assigning an initial PreviousElo value to be the same as CurrentElo. This change will help in tracking the initial Elo value before any changes are made to a player's Elo which is useful for stat tracking.