Add authorization views and update icons
Added authorization views to control access to confirm result buttons based on user role. Additionally, the icons in the site navigation have been updated to use Blazorise icons instead of open-iconic icons. This was done to create a more cohesive branding across the site and to improve aesthetics. Also, the CSS class for icon spacing was added to style tag for better readability.