Enhance player ranking UI and data representation
This commit enhances the UI of the player ranking page by encapsulating player details in PlayerCard component for better maintainability and readability. Also, additional information like the 'LastEloGained' and 'MatchesPlayed' has been added to the PlayerRankingResponse. This gives the users a better understanding of player's performance. Furthermore, players with more than one Elo history are now considered in the ranking list to ensure that the rankings are based on sustained performance. Minor changes are implemented in type configurations to maintain code efficiency. Also, a new package 'Blazor.Flags' has been added for future use.