Update user identification method in controllers
The previously used method of identifying the user in the controllers relied on a simple 'userId'. To ensure secure identification and match it with the ID provided by the authentication middleware, the code has been modified to translate 'userId' into a 'Guid'. This was done across all controllers wherever the 'userId' had been previously used for identification. The Player creation process has also been updated to include this change.