Add health checks for application database
This commit includes the addition of the health check package for Entity Framework Core and the necessary code to implement these checks for the application database. Health checks provide a way to monitor the status and health of an application. They were added to the Startup file and will be accessible via '/hc' for full checks and '/hc-lite' for lighter ones. These changes will enhance the project's robustness, making it easier to identify and solve possible failures.