Add player detail endpoint and view
Add a new endpoint to retrieve player detail and a corresponding view to display this info. This will allow users to get more info about a player by selecting them in the ranking list. The PlayerRankingResponse model was renamed and reused for simplicity, refactoring existing components where it was needed. Also removed an unused link in MainLayout.