Replace LoadingSpinner with PlayerCardSkeleton for better UX
The LoadingSpinner component was replaced with a skeleton screen, using the PlayerCardSkeleton component, to improve the user experience during data loading. Instead of showing a spinner, now, skeleton cards are displayed, providing a sense of content structure and layout, which helps to mitigate the perception of delay. This method is considered as best practice, especially for pages with potentially high loading times.

Also, stylistic CSS and necessary HTML are added for the new skeleton component.