"Refactor PlayerEloProgress and PlayerController for improved data handling"
This commit primarily addresses two changes. Firstly in PlayerEloProgress.razor, the check for "_eloHistory != null" has been removed from the if condition within OnAfterRenderAsync(). This was done to ensure that the rest of the code in the block executes regardless of whether _eloHistory is null or not. A new block was also added to handle scenarios where the _eloHistory is null but there's OtherPlayerEloHistory.

Secondly, in PlayerController.cs, the sorting order for the EloHistories list has been changed from descending order based on Elo to ascending order based on ChangeDate. This will ensure that we have chronologically sorted data which is more