"Update ranking query in PlayerController"
Removed the filtering of players with more than one EloHistory from the GetRanking method in PlayerController.
This change was made to ensure that all players are included in the ranking, regardless of their number of EloHistories.
Changed the ordering of players by first considering the number of matches played before applying ELO ranking, to give priority to active players.