Add player rank to player detail view
A new property 'Rank' has been added to the PlayerDetailResponse object, it is displayed on the player card component in the clientside. The server-side PlayerController has been updated to calculate the rank based on numbers of matches played and Elo score. This feature was added to provide more detailed information about a player's performance and ranking in the game.