Remove divider, improve chart design
The divider was removed from the Index.razor page for aesthetic reasons.
This makes the UI cleaner and more streamlined.

In the PlayerEloProgress.razor component, classes for border and chart card styles were added for making the chart more visually appealing and user-friendly. This includes adding space around the chart, a margin at the bottom, and a subtle box-shadow for depth. The 'if' conditions for 'OtherPlayerEloHistory' were also wrapped inside the div element for cleaner code organization.