Add player profile navigation from user image
This commit introduces a clickable user image that navigates to the associated player's profile. It includes updates to the UserDTO to maintain consistency of the user information across components. A PlayerId property has been added to the UserDto and used in the UserImage component to navigate to the player detail page when the user image is clicked. Also, the UserImage.razor file has been updated to include NavigationManager and an onclick event to enable this navigation. The GetMyUserQueryHandler has been modified accordingly to return the new PlayerId.