Update PlayerController to include players with single Elo history
In the PlayerController.cs, the LINQ query in method was modified. Players used to be filtered based on whether their EloHistories count is more than one. This led to exclusion of players who have only played one game. This has been changed to include players even if they have a single Elo history by replacing "Count > 1" with "Any()", thus enabling these players to be part of the query results.