Add authentication check to update ELO history label
The changes in PlayerEloProgress.razor ensure that the label displayed on the ELO history chart corresponds to the authentication state of the user. A new boolean "isAuth" has been introduced to maintain the state of authentication. If the user is authenticated, the label of the chart is updated to "Your ELO history", otherwise, it remains "His ELO history". The method GetLineChartOtherPlayerDataset has also been made asynchronous to accommodate this change. Additionally, an unnecessary divider in PlayerDetail.razor has been removed for cleaner UI.