Add Elo change reason for players
Added two new columns in Players and EloHistories for storing the reason of the Elo change.
The change was necessary in order to keep track of actions that affect a player's Elo ranking.
This includes instances of adjustments made by an admin and match results.
This information is essential for dispute resolution and auditing purposes.