Add "last matches" functionality to player detail page
An HTTP Get method has been added to the PlayerController to fetch the last five matches a player has been involved in. The fetched matches contain data like the match ID, start time, duration, average Elo, and Elo change for the player.
This information was then used to create the 'LastMatchesResponse' class, a new LastMatches component on Razor which shows this information on the frontend, and the 'LastMatches' was added to PlayerDetail to display this new functionality.

This commit was necessary to provide users a way to quickly see a player's recent performance without having to extensively search through all their matches.