Add a 'Winner' column to the 'Match' table and associate it with team
A new 'Winner' column has been added to the 'Match' table to keep track of the winning team. This attribute has been mapped to the 'Team' enumerated type for consistency with other related fields. Now, after each match completion, the winner will be calculated based on set scores and recorded in the 'Match' table, making the tracking of outcomes more explicit. It allows for more precise data analysis and enables to easily derive match statistics.