Refactor ShowResults component condition check
Removed a redundant condition check in the ShowResults.razor component of the Match feature. The previous version checked if Match.RequesterIsTheCreator or if there are any sett in a Match before rendering the "match-set-results". The check was eliminated because regardless, the "match-set-results" needs to be rendered. This simplifies the code and makes it more readable.