Refactor authorization to use Auth0 instead of built-in Identity
Removed the CustomUserFactory class since it's no longer needed with the usage of Auth0. Modified the behavior of the Application's Identity Server to authenticate through Auth0 instead of using the default system. Introduced the JWT Bearer scheme in server-side Startup to use Auth0 for authentication. Removed unused components and methods related to the old authentication system in the Client project, and added necessary components to handle Auth0 authentication. The change was necessary to meet the new requirements of the project regarding user authentication, which requires the use of Auth0's services. It improves the security and simplifies the codebase by decoupling the authentication system from the application's server.