"Enhance email confirmation and user role system"
In this commit, some primary changes were made to improve user confirmation and role assignment processes. An email service was introduced and implemented so that emails are sent to users right after the registration process. In this process, a confirmation link is shared with the user. A newly registered user is also assigned the "USER" role immediately after registration. In the code, we have refactored a bit to increase efficiency. Instead of keeping unencoded and encoded values for the same variable, we have directly assigned an encoded value. In the ConfirmEmail view, we have revised the logic to enhance readability and performance. We have adopted a ternary operator. If the email confirmation process gets successful a "Thank You" message will be shown. Otherwise, an error message will be displayed. Further, UI part was enhanced by changing few button colors and adding a button to navigate back to the app.