Refactor email service usage in account authentication
This commit incorporates the EmailService into the authentication process for user accounts. Initially, email setup was done via the IEmailSender, which was subsequently used for sending mails. However, due to an upgrade in the service setup, IEmailService is now implemented across the ExternalLogin, Register, and ForgotPassword models. This change simplifies the email handling process and also enhances the service's testability. Furthermore, adjustments were made to the Execute method's parameters within the IEmailService interface and EmailService class to accommodate these changes.