Add role assignment to external login flow
A change was made in the external login process to assign a new user to a default role. A new 'Player' is created with default values upon new user registration. The user is given the "User" role and a Player record is created in the database with the initial Elo ranking of 1500. This addition is necessary for proper role assignment and tracking of player statistics from the moment of account creation. The 'context' parameter has been added for making database operations.