Add loading spinner to register and login buttons
The registration and login buttons on the Register, Login, and ExternalLogin pages have been updated. The changes enable a loading spinner to be displayed after clicking the submission button, providing visual feedback that the process is underway. This enhances UX by informing users when a process is ongoing especially if there's a long wait time due to slow network speeds.

In addition to this, the Google login/register buttons have been updated to include a `click` event listener that disables the button and displays a spinner immediately after click. This prevents multiple submit requests and provides immediate feedback to the user.

A form validity check is also introduced on the registration form to enable/disable the submit button based on the form's validity at any moment.