Resize and reposition match status tags
The CSS properties for the status tags of matches in MyMatches page have been resized and repositioned. The '.finished-stamp', '.searching-stamp', and '.pending-stamp' classes now have a smaller font size, width, and are placed closer to the bottom-right edge of the container. This change improves placement and visibility of these elements on smaller screen sizes and provides a more streamlined look.