Remove local time conversion for timestamps
The time-conversion to local time was removed from various parts of the application. This change was necessary as it was causing issues with time discrepancy when viewed from various timezones. Now, the timestamps for match start date and suggestion create date are shown in the server's set universal time, eliminating the time discrepancy between different client views.